social_media_api/requirements.txt
Django>=3.2.0
djangorestframework>=3.12.0
Pillow>=8.0.0
